refactor(preconfigureJobs): Move KubernetesPreconfiguredJobProperties to orca-api
